Horbyon Studios/ClosetdaJacy/Painel-Admin/backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
pymysql==1.1.1
bcrypt==4.2.0
python-multipart==0.0.9
python-jose[cryptography]==3.5.0
python-dotenv==1.0.1

